+++ EnemiesScannerMod/ModVariables.cs
@@ -13,6 +13,9 @@ namespace EnemiesScannerMod
         public AudioClip RadarAlertSound { get; set; }
         public AudioClip OverheatedSound { get; set; }
         public AudioClip RebootedSound { get; set; }
+        public AudioClip NoPowerSound { get; set; }
+        public AudioClip TurnOnSound { get; set; }
+        public AudioClip TurnOffSound { get; set; }
         
         public GameObject ModNetworkManagerGameObject { get; set; }